(refactor) idesc_table: Merge headers
Merging src/include/kernel/task/idesc.h with
        src/include/kernel/task/resource/idesc.h into second one.

This headers were used together often and _always_ using the first
one implicitly made dependency on corresponding resources, so this
merge is necessary to make code more clear.
