+++ src/compat/posix/fs/fcntl.c
@@ -9,7 +9,7 @@
 
 #include <fs/idesc.h>
 #include <fs/index_descriptor.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 int fcntl(int fd, int cmd, ...) {
 	int ret;
