+++ src/compat/posix/fs/fstat.c
@@ -12,7 +12,7 @@
 
 #include <fs/idesc.h>
 #include <fs/index_descriptor.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 int fstat(int fd, struct stat *buff) {
 	int ret;
