+++ src/compat/posix/fs/ftruncate.c
@@ -11,7 +11,7 @@
 
 #include <fs/index_descriptor.h>
 #include <fs/kfile.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 extern const struct idesc_ops idesc_file_ops;
