+++ src/compat/posix/fs/write.c
@@ -11,7 +11,7 @@
 
 #include <fs/index_descriptor.h>
 #include <fs/idesc.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 ssize_t write(int fd, const void *buf, size_t nbyte) {
 	ssize_t ret;
