+++ src/compat/posix/idx/dup.c
@@ -12,7 +12,7 @@
 #include <fcntl.h>
 
 #include <kernel/task.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 int dup(int flides) {
 	return fcntl(flides, F_DUPFD, 0);
