+++ src/compat/posix/idx/idesc_pipe.c
@@ -19,7 +19,6 @@
 #include <framework/mod/options.h>
 #include <kernel/thread/sync/mutex.h>
 #include <kernel/task.h>
-#include <kernel/task/idesc_table.h>
 #include <kernel/task/resource/idesc_table.h>
 #include <fs/idesc.h>
 #include <fs/idesc_event.h>
