+++ src/compat/posix/idx/select.c
@@ -20,7 +20,7 @@
 #include <kernel/sched.h>
 #include <kernel/task.h>
 
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 #include <fs/index_descriptor.h>
 #include <fs/idesc.h>
 #include <fs/idesc_event.h>
