+++ src/init/diag_fd.c
@@ -9,7 +9,7 @@
 #include <termios.h>
 #include <drivers/diag.h>
 #include <kernel/task.h>
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 
 #include <fs/idesc.h>
