+++ src/kernel/task/idesc/idesc_table.c
@@ -15,7 +15,7 @@
 #include <fs/idesc.h>
 #include <kernel/task.h>
 
-#include <kernel/task/idesc_table.h>
+#include <kernel/task/resource/idesc_table.h>
 #include <util/array.h>
 #include <util/indexator.h>
