+++ src/kernel/task/idesc/index_descriptor.c
@@ -9,7 +9,6 @@
 
 #include <kernel/task.h>
 
-#include <kernel/task/idesc_table.h>
 #include <kernel/task/resource/idesc_table.h>
 #include <fs/idesc.h>
