+++ src/net/socket/socket_desc.c
@@ -7,7 +7,6 @@
 #include <stddef.h>
 
 #include <kernel/task.h>
-#include <kernel/task/idesc_table.h>
 #include <kernel/task/resource/idesc_table.h>
 #include <net/socket/socket_desc.h>
 #include <fs/idesc.h>
